Fixes for Authorization and error handlers

+++ blog.py
@@ -126,6 +126,7 @@ class PostPage(BlogHandler):
                     comments=comments, error=error)
 
     def post(self, post_id):
+
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         post = db.get(key)
 
@@ -210,6 +211,11 @@ class DeletePost(BlogHandler):
         if self.user:
             key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
+            
+            if not post:
+                self.redirect("/")
+                return
+
             if post.user_id == self.user.key().id():
                 post.delete()
                 self.redirect("/?deleted_post_id="+post_id)
@@ -226,6 +232,11 @@ class EditPost(BlogHandler):
         if self.user:
             key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             post = db.get(key)
+           
+            if not post:
+                self.redirect("/")
+                return
+
             if post.user_id == self.user.key().id():
                 self.render("editpost.html", subject=post.subject,
                             content=post.content)
@@ -242,7 +253,7 @@ class EditPost(BlogHandler):
         """
         if not self.user:
             self.redirect('/blog')
-
+    
         subject = self.request.get('subject')
         content = self.request.get('content')
 
@@ -266,6 +277,11 @@ class DeleteComment(BlogHandler):
             key = db.Key.from_path('Comment', int(comment_id),
                                    parent=blog_key())
             c = db.get(key)
+
+            if not c:
+                self.redirect("/")
+                return
+                
             if c.user_id == self.user.key().id():
                 c.delete()
                 self.redirect("/blog/"+post_id+"?deleted_comment_id=" +
@@ -284,6 +300,11 @@ class EditComment(BlogHandler):
             key = db.Key.from_path('Comment', int(comment_id),
                                    parent=blog_key())
             c = db.get(key)
+           
+            if not c:
+                self.redirect("/")
+                return
+                
             if c.user_id == self.user.key().id():
                 self.render("editcomment.html", comment=c.comment)
             else:
